pyproject.toml
[project]
authors = [
    {name = "nayutah", email = "dullboy@apecloud.com"},
]
requires-python = "<3.13,>=3.10.0"
dependencies = [
    "langchain<1.0.0,>=0.3.23",
    "langchain-huggingface<1.0.0,>=0.1.2",
    "llama-index<1.0.0,>=0.12.28",
    "uvicorn[standard]<1.0.0,>=0.23.1",
    "django==5.0.14",
    "pydantic[email]>=2.10.3,<3.0.0",
    "cryptography<42.0.0,>=41.0.1",
    "python-dotenv<2.0.0,>=1.0.0",
    "django-environ<1.0.0,>=0.10.0",
    "django-ninja<2.0.0,>=1.4.0",
    "auth0-python<5.0.0,>=4.2.0",
    "celery<6.0.0,>=5.3.1",
    "channels<5.0.0,>=4.0.0",
    "py7zr<1.0.0,>=0.20.8",
    "rarfile<5.0,>=4.1",
    "qdrant-client<2.0.0,>=1.3.0; python_version < \"3.12\" and python_full_version >= \"3.8.1\"",
    "redis<5.0.0,>=4.6.0",
    "arrow<2.0.0,>=1.2.3",
    "func-timeout<5.0.0,>=4.3.5",
    "terminal<1.0.0,>=0.4.0",
    "types-chardet<6.0.0.0,>=5.0.4.6",
    "psycopg2-binary<3.0.0,>=2.9.6",
    "watchfiles<1.0.0,>=0.19.0",
    "flower<3.0.0,>=2.0.0",
    "django-celery-beat<3.0.0,>=2.5.0",
    "boto3>=1.26.165,<2.0.0",
    "oss2<3.0.0,>=2.18.0",
    "django-cors-headers<5.0.0,>=4.1.0",
    "socksio<2.0.0,>=1.0.0",
    "django-prometheus<3.0.0,>=2.3.1",
    "whitenoise<7.0.0,>=6.5.0",
    "types-requests<3.0.0.0,>=2.31.0.2",
    "elasticsearch<9.0.0,>=8.9.0",
    "gevent<24.0.0,>=23.9.1",
    "requests<3.0.0,>=2.31.0",
    "qianfan<1.0.0.0,>=0.4.12.3",
    "pillow<11.0.0,>=10.0.0",
    "openai<2.0.0,>=1.71.0",
    "dashscope<2.0.0,>=1.23.1",
    "gitpython<4.0.0,>=3.1.43",
    "llama-index-embeddings-google<1.0.0,>=0.3.1",
    "llama-index-embeddings-langchain<1.0.0,>=0.3.0",
    "llama-index-vector-stores-opensearch<1.0.0,>=0.5.2",
    "llama-index-vector-stores-milvus<1.0.0,>=0.7.2",
    "llama-index-vector-stores-qdrant<1.0.0,>=0.6.0",
    "llama-index-vector-stores-weaviate<2.0.0,>=1.3.1",
    "numpy<2.0.0,>=1.26.4",
    "lightrag-hku<2.0.0,>=1.3.7",
    "pipmaster<1.0.0,>=0.5.4",
    "graspologic<4.0.0,>=3.4.1",
    "pytablewriter<2.0.0,>=1.2.1",
    "litellm>=v1.68.0",
    "requests",
    "asyncpg>=0.30.0",
    "markitdown[all]>=0.1.1",
    "markdown-it-py[linkify]>=3.0.0",
    "nano-vectordb>=0.0.4.3",
]
name = "aperag"
version = "0.1.0"
description = "A LLM based tool to support Text2SQL, Chatbot, KubeBlocks pilot and troubleshooting"
readme = "README.md"

[tool.uv]
# index-url = "https://mirrors.aliyun.com/pypi/simple/"
prerelease = "allow"

[project.optional-dependencies]
all = [
    "pymilvus<3.0.0,>=2.2.12",
    "weaviate-client<5.0.0,>=4.5.7",
]
lightrag-dev = [
    "nano-vectordb"
]
evaluation = [
    "ragas"
]

[project.scripts]
test_local_path_embedding = "readers.test_local_path_embedding:start"

[dependency-groups]
model = [
    "torch<3.0.0,>=2.6.0",
    "text2vec<2.0.0,>=1.3.3",
    "flagembedding<2.0.0,>=1.3.4",
    "sentence-transformers<5.0.0,>=4.0.2",
    "transformers==4.49.0",
]
dev = [
    "ruff<1.0.0,>=0.11.2",
    "types-toml<1.0.0.0,>=0.10.8.1",
    "types-redis<5.0.0.0,>=4.3.21.6",
    "types-chardet<6.0.0.0,>=5.0.4.6",
    "vulture<3.0,>=2.14",
    "deptry<1.0.0,>=0.23.0",
    "mypy<2.0.0,>=1.4.1",
    "datamodel-code-generator>=0.30.0",
]
test = [
    "tabulate<1.0.0,>=0.9.0",
    "pytablewriter<2.0.0,>=1.0.0",
    "pytest",
    "pytest_asyncio",
    "moto",
]

[tool.ruff]
lint.select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
]

lint.ignore = [
    "E501", # line too long
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".github",
    "models",
    "playground",
    "aperag/migrations",
    "aperag/schema/view_models.py",
    "vectorstore/test_connector.py",
    "readers/test_local_path_embedding.py",
]

# Same as Black.
line-length = 120

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "False"
exclude = [
    "playground",
    "services/text2SQL",
    ".github",
    "models",
    "playground",
    "aperag/migrations",
    "vectorstore/test_connector.py",
    "readers/test_local_path_embedding.py",
]

[tool.setuptools]
packages = ["aperag", "config"]
